requirements.txt
virtualenv==15.1.0
numpy==1.12.0
numpy==1.12.0
pandas==0.19.2
scipy==0.18.1
scikit-learn==0.18.1
scikit-image==0.12.3
opencv-python==3.2.0.6
numpy==1.12.0
scipy==0.18.1
pandas==0.19.2
scikit-learn==0.18.1
scikit-image==0.12.3
dlib==19.1.0
PyOpenGL==3.1.0
PyOpenGL-accelerate==3.1.0
pygame==1.9.3
pyglet==1.2.4
flake8==3.3.0
Pillow==4.0.0
